T2T NDEF Tag : Wrong ndef msg length(more than tag capacity)is leading to endless loop
When reading a NDEF T2T tag where the ndef msg length is exceeding the maximum capacity of tag
the stack is going in an endless loop. The phone is trying to read the full length but when the
read pointer is out of the tag capacity, the tag will not respond anymore. This leads to
disconnection and reconnection. The phone will read the full length again and again. Expected
behavior is to start the presence check soon after the ndef read procedure is done. We were
wrongly setting the NDEF detection time out although we have done it.