NFC Initialization failed
Added logic to add HCI callback for HOST. As per NCI2.0 changes, nfcee
discovery notification for DH-NFCEE are not received. Only notifications
for off-host NFCEE are received, due to which, during initialization,
HCI callback are not getting registered for DH-NFCEE although its
available all the time. This case is handled for standalone controllers.