Fix for screen state update while in mPOS Reader Mode
If transaction is ongoing and screenState request received, store only
if it is not same as current screenState.
During stopEtsiConfig, enableDiscovery will not be sent until
mPOS state in ETSI_READER_STOP. This causes corruption in discovery parameters
as same time pending screen state commands will be sent.
So set mPOS state to ETSI_READER_STOP and then call enableDiscovery.