eSE Access Interface - NFC extras==>Synchronization between OMAPI & NFC extras -SSG
Description:
isWiredEnable flag is used to control the open of SE through Omapi as well as Extras.
Extras transcieve is handle by pid it will not allow until unless open of Extras is not called.
Omapi transmit is not called until unless openSecureelement is succes.

Test:
Open wired on OMPAI, simaltaneously open on extras, service will return the SE is open
Open wired on Extras, simaltaneously open on Ompai, service will return the SE is open