Polling mode gets disabled in between while tapping Beijing, Chang'an and Cash Bee card iteratively.
The p2p_prio_logic_multiprotocol() functions is called as a separate task from multiple places in the code.
The issue comes when more than one task of this function is in life and runs simultaneously.
The function's code has been kept under Critical Section i.e. used Mutex to achieve the Synchronization.