(Q_AOSP) Correct routing behavior in screen off.
Change is already part of NXP_EXTNS. Now it is added
as default AOSP change.
In screen off, if the DH-NFCEE is considered as
disabled, the corresponding entry in the routing
table is ignored (NCI specification chap 6.3).
Therefore, in screen off a contactless transaction
will be routed to offhost (by protocol/technology
routing) even if in the entry the AID was blocked
for all power state not supported by the DH.

To prevent this erratic behavior and to conform
to GSMA TS26_NFC_REQ_147, the DH-NFCEE should be
considered enabled in screen off.