Fix broken conflict resolution logic
As per MultiSE change, if any prefix AID found which has
conflict with other AIDs, Its route location is set to Host
only if corresponding service is of Payment category.
else it will consider route location set in offHostSE.
This causes prefix AID to route to offHost(if AID corresponding
to offHost service) which will not be considered for
conflict resolution.
Fix:This logic is reverted to set route location to host directly
if any prefix AID available to be resolved