Fix for NFC died observed during Mifare Classic and Mifare Desfire tag read .
MIFARE DESFIRE and MIFARE CLASSIC tags are tapped with taginfo application.
Sometimes duirng reselect failed , Mifare desfire will be detected as mifare classic
because of start rf discovery. Therefore , the protocol will be set
to desfire but libnfc-nci will be updated to MIFARE CLASSIC.Because of this,
mifare classic commands are sent from libnfc-nci and FW doesnt respond.NFC service
died is observed because of this.