{QSSI} Fix duplicate routing update during dual UICC switch
During dualUicc switch if there are AIDs, then commitRouting is called after
AIDs are sent.Then, as part of dualUICC switch, again commitRouting is called
(required in case if no AIDs available). This causes duplicate routing table update.
To avoid this, an existing flag is used, which will be set if commitRouting
is called as part of AID routing. If this flag is set,the do not call
commit routing again.