Set default power state when NFC is enabled
If NFC is turned off and on the same power state,
DH doesn't send CORE_SET_POWER_SUB_STATE_CMD to NFCC
because previous power state in DH is same as the new one.

But, actually, default power state in NFCC is Switched On.
Therefore, if NFC is enabled on the "locked" state,
secure NFC doesn't work until power state is changed.

This patch sets previous power state of Switched On-Unlocked
so that CORE_SET_POWER_SUB_STATE_CMD is correctly updated
when NFC is enabled.