{Q_AOSP}:Give room to register NFC adapter for payment service
When NFC service is created, it binds to a default payment
service. Then the payment service fails to get NfcAdapter
instance if the service process is created before NFC is
registered to service manager.

On the other hand, CardEmulationManager instance should be
created before service registeration in order to ensure
providing CardEmulation interface.

Therefore, delay calling bindService for payment service to
ensure that NFC service is ready.