Beam: Show beam UI even if taking screenshot is failed
There is confusing behavior when beam is started
by ACTION_SEND. If taking screenshot fails due to
FLAG_SECURE or something, beam UI is not shown
nevertheless beam is ready to send a content.

On the other hand, UI is shown in that case when
beam is started by tap. The same logic should be
applied to the case of ACTION_SEND as well.