Do not set mBluetoothEnabledByNfc when BT state changed
mBluetothEnabledByNfc is used to decide whether to disable BT
when finish BT handover.

Basically the value is set to false in onComplete() method and
it's enough to control the state. However it's changed also in
BroadcastReceiver when BT state changed to disabled.

It's been redundant and if the broadcast intent delays,
then it would cause state mismatch.