{AOSP}: Disconnect tag during screen-off only if tag operation is active
During screen state update to screenoff, Tag is forcibly
disconnected. But this needs to be done only if tag operation is
active and not when CE/P2P is ongoing.

In case of CE-eSE, Information about technology and other parameters
are not available during activation. But EE_DIRECT_RF interface is
selected. This is used to identify device is in listen mode.