Fix No popup window for conflict aid selection
As part of aligning MW code to AOSP, new changes are added
for calculation of route location.
with this change, if AID is of payment category, then only
it will be updated with route location host, else it will take
route location of the AID service itself.
This deviates from previous nxp change where in case
of non default single service, by default route location is
update to Host for both payment and non payment AIDs.
which allows popup when conflicting AID selected.
To keep this behavior as it is, AOSP change is reverted
and by default onhost will be set to the AID.