Fix to non-proprietary tag reading failure.
If DEACTIVATE to IDLE fails for defined time, then MW issues DEACTIVATE
to DISCOVER. Then, after checking the type and mode of the previously
activated tag, it decides either it is non-std Tag or not.

For this, MW shall save previously activated tag data(type and mode)
and retain after receiving the DEACTIVATE to DISCOVER rsp.