Fix missing mServiceBound state update
Failure to update this variable will cause mService and mServiceName not
to be nullified if onHostEmulationDeactivated() is called right after
onServiceConnected() is called. onHostEmulationDeactivated() calls
unbindServiceIfNeededLocked() which uses mServiceBound to unbind the
service. Since mServiceBound was not updated to the correct value,
unbindServiceIfNeededLocked() will never be able to unbind the
previously bound service.