Add required flag to registerReceiver calls in nfc/beam
Android T adds support to allow a runtime receiver to be registered as
not exported, but to ensure apps can take advantage of this, calls to
registerReceiver must specify a flag indicating whether the receiver
should be exported for apps targeting T+ that are registering for
unprotected broadcasts. This commit adds the RECEIVER_EXPORTED
flag to all calls to registerReceiver under nfc/beam/ as the receivers
are either guarded by a signature|priv permission or are only
accessible via testing (ACTION_CLOSE_SYSTEM_DIALOGS).

Bug: 161145287
Test: Build

Change-Id: I7013da28b449dcb9d3750927cd5d71afaf8d3353