Fix Nfc Service stuck issue.
When Nfc Service Off is triggered in parallel with Unrecoverable
error handling, The issue comes when the disableInternal AsynchTask
waits to lock the NfcService object while the same is locked by the
service message handler while updating the LMRT as part of error recovery.

In this case, disableInternal already terminates  the NFA_Task.
So, when the Service message Handler posts an event to update routing
entries to the NFA_Task() it is lost, not served. Thus, Service
message handler waits for infinite time causing deadlock to the
DisableInternal Async Task.

Change-Id: Ie8248d569ff744939c930fc490c2099922f2c495